requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy
pandas
matplotlib
tqdm
scipy
soundfile
scikit-learn
pyyaml
